Actualizando y validando.

+++ BackEnd/src/main/java/pe/edu/upc/backend/entities/Pais.java
@@ -1,7 +1,5 @@
 package pe.edu.upc.backend.entities;
-
 import jakarta.persistence.*;
-
 @Entity
 @Table(name = "Pais")
 public class Pais {
@@ -23,12 +21,9 @@ public class Pais {
     @Column(name = "poblacionPais", nullable = false)
     private Long poblacionPais;
     @Column(name = "superficiePais", nullable = false)
-    private int superficiePais;
-
-    public Pais() {
-    }
-
-    public Pais(int idPais, String nombrePais, String codigoIsoPais, String continentePais, String capitalPais, String monedaPais, String idiomaOficialPais, Long poblacionPais, int superficiePais) {
+    private Long superficiePais;
+    public Pais() {}
+    public Pais(int idPais, String nombrePais, String codigoIsoPais, String continentePais, String capitalPais, String monedaPais, String idiomaOficialPais, Long poblacionPais, Long superficiePais) {
         this.idPais = idPais;
         this.nombrePais = nombrePais;
         this.codigoIsoPais = codigoIsoPais;
@@ -39,76 +34,56 @@ public class Pais {
         this.poblacionPais = poblacionPais;
         this.superficiePais = superficiePais;
     }
-
     public int getIdPais() {
         return idPais;
     }
-
     public void setIdPais(int idPais) {
         this.idPais = idPais;
     }
-
     public String getNombrePais() {
         return nombrePais;
     }
-
     public void setNombrePais(String nombrePais) {
         this.nombrePais = nombrePais;
     }
-
     public String getCodigoIsoPais() {
         return codigoIsoPais;
     }
-
     public void setCodigoIsoPais(String codigoIsoPais) {
         this.codigoIsoPais = codigoIsoPais;
     }
-
     public String getContinentePais() {
         return continentePais;
     }
-
     public void setContinentePais(String continentePais) {
         this.continentePais = continentePais;
     }
-
     public String getCapitalPais() {
         return capitalPais;
     }
-
     public void setCapitalPais(String capitalPais) {
         this.capitalPais = capitalPais;
     }
-
     public String getMonedaPais() {
         return monedaPais;
     }
-
     public void setMonedaPais(String monedaPais) {
         this.monedaPais = monedaPais;
     }
-
     public String getIdiomaOficialPais() {
         return idiomaOficialPais;
     }
-
     public void setIdiomaOficialPais(String idiomaOficialPais) {
         this.idiomaOficialPais = idiomaOficialPais;
     }
-
-    public Long getPoblacionPais() {
-        return poblacionPais;
-    }
-
+    public Long getPoblacionPais() { return poblacionPais; }
     public void setPoblacionPais(Long poblacionPais) {
         this.poblacionPais = poblacionPais;
     }
-
-    public int getSuperficiePais() {
+    public Long getSuperficiePais() {
         return superficiePais;
     }
-
-    public void setSuperficiePais(int superficiePais) {
+    public void setSuperficiePais(Long superficiePais) {
         this.superficiePais = superficiePais;
     }
 }
